review 2:Notify

+++ database/notify.go
@@ -4,7 +4,7 @@ import "time"
 
 // Notify 通知
 type Notify struct {
-	NotifyID  int    `gorm:"primary_key;AUTO_INCREMENT"`
+	NotifyID  uint64 `gorm:"primary_key;AUTO_INCREMENT"`
 	Content   string `gorm:"type:TEXT;not null"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
